更新用户模型，将用户 ID 类型从 uint64 修改为 uint，调整 ID 生成函数以返回 uint 类型，同时修复 JWT 生成时用户 ID 的转换方式。

+++ example/model/user.go
@@ -9,7 +9,7 @@ import (
 
 type User struct {
 	gorm.Model
-	Uid   uint64 `gorm:"unique"`
+	Uid   uint   `gorm:"unique"`
 	Email string `gorm:"unique"`
 }
